requirements.txt
streamlit
pandas
google-auth
python-dotenv
gspread
gspread-dataframe
matplotlib